Created data types for Resources.
FireResource is an interface with a location and resource name.
It is intended to be implemented on `FireFile`s, `FireDir`s, `FireStruct`s, `FireFunction`s, etc.

ResourceLocation is a format for resource locations such as `project::folder::file::struct::method`.

ResourceName is a string wrapper that only permits valid resource names. (Match the regex `^[a-zA-Z][a-zA-Z0-9_]*$`)