Created data types for Resources.
FireResource is a class with a resource name and maybe a parent.
Intended for `FireFunction`, etc

FireContainerResource is a FireResource but also contains children resources.
Intended for `FireFile`, `FireFolder`, `FireStruct`, `FireEnum`, etc.

ResourceLocation is a format for resource locations such as `project::folder::file::struct::method`.

ResourceName is a string wrapper that only permits valid resource names. (Match the regex `^[a-zA-Z][a-zA-Z0-9_]*$`)